requirements.txt
transformers
torchmetrics
datasets
evaluate
PyYAML